refactor: Streamline ClickOptions and ClickOptionsBuilder access
The `page` module within `handler` is now private. `ClickOptions` and
`ClickOptionsBuilder` are re-exported directly from `crate::handler`
and `crate::` for simpler imports and a cleaner public API surface.
