+++ src/page.rs
@@ -29,11 +29,11 @@ use chromiumoxide_types::*;
 use crate::auth::Credentials;
 use crate::element::Element;
 use crate::error::{CdpError, Result};
+use crate::handler::ClickOptions;
 use crate::handler::PageInner;
 use crate::handler::commandfuture::CommandFuture;
 use crate::handler::domworld::DOMWorldKind;
 use crate::handler::httpfuture::HttpFuture;
-use crate::handler::page::ClickOptions;
 use crate::handler::target::{GetName, GetParent, GetUrl, TargetMessage};
 use crate::js::{Evaluation, EvaluationResult};
 use crate::layout::Point;
